services/api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
pydantic